+++ src/app/features/events/events.component.ts
@@ -28,7 +28,14 @@ export class EventsComponent implements OnInit, OnDestroy {
   @ViewChild(MatSort, {static: false}) sort: MatSort;
 
   public viewMode = 'cards';
+  public statusFiltered = [
+    Status.Visible,
+    Status.Editing,
+    Status.Blocked,
+    Status.Deleted
+  ];
   public events: IEvent[];
+  public EVENTS_BACKUP: IEvent[];
   public dataSource: MatTableDataSource<IEvent> = new MatTableDataSource();
   displayedColumns: string[] = [
       'status', 'id', 'timestamp',
@@ -68,13 +75,14 @@ export class EventsComponent implements OnInit, OnDestroy {
           }))
         )
         .subscribe( (events: IEvent[]) => {
-        this.events = events;
-        this.dataSource = new MatTableDataSource(this.events);
-        this.spinnerSvc.hide();
+            this.events = events;
+            this.EVENTS_BACKUP = this.events;
+            this.dataSource = new MatTableDataSource(this.events);
+            this.spinnerSvc.hide();
 
-        this.dataSource.paginator = this.paginator;
-        this.dataSource.sort = this.sort;
-    });
+            this.dataSource.paginator = this.paginator;
+            this.dataSource.sort = this.sort;
+        });
 
     this.listOfObservers.push(subs2$);
   }
@@ -132,6 +140,14 @@ export class EventsComponent implements OnInit, OnDestroy {
 
   setViewMode(mode: string): void {
     this.viewMode = mode;
+  }
+
+  setStatusFiltered(filtered: Status[]): void {
+    this.statusFiltered = filtered;
+    console.log(`${this.statusFiltered}`);
+    this.events = this.EVENTS_BACKUP.filter(notice =>
+      this.statusFiltered.includes(notice.status)
+    );
   }
 
   private formatSocialInfo(info: string): string {